requirements.txt
# Flask Job Scheduler
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-SocketIO>=5.3.0
gevent>=23.0.0
gevent-websocket>=0.10.1

# Scheduler
APScheduler>=3.10.0

# Remote Execution
paramiko>=3.0.0
pywinrm>=0.4.3
requests>=2.28.0

# Ansible (optional - user may install separately)
# ansible>=2.14.0

# Utilities
python-dotenv>=1.0.0
croniter>=1.3.0
